+++ libraries/HC4051/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2024-04-03
+- fix ghost channels when using for OUTPUT
+  - add disable/enable in setChannel()
+- update setChannel() performance.
+- update readme.md
+- update GitHub actions
+- minor edits
+
+----
+
 ## [0.1.1] - 2023-11-04
 - update readme.md